% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{air_dens}
\alias{air_dens}
\title{Find air density based on Temp, Humidity, and Pressure}
\usage{
air_dens(T_C, RH, P_mbar)
}
\arguments{
\item{T_C}{Air temperature (deg. C)}

\item{RH}{Relative Humidity (0-100\%)}

\item{P_mbar}{Air pressure (mbar)}
}
\description{
This method calculates air density (kg m-3) based on Air temperature (C)
and Relative Humidity (\%) using some crazy ass equations I found
on the internet.
https://www.calctool.org/atmospheric-thermodynamics/air-density
}
\note{
There should be a note here
}
\examples{
#there will be an exmaple here
}
