% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drag}
\alias{drag}
\title{Profile method for calculating wind drag}
\usage{
drag(Az, Uz, z = NULL, Cd = 0.7, rho = 1.204)
}
\arguments{
\item{Az}{vector representing cross sectional area (m^2) of each 1 m tree/canopy segment}

\item{Uz}{vector representing wind speed (m/s) at each 1 m tree/canopy segment}

\item{z}{vector of heights of tree, 1m segements assumed if NULL}

\item{Cd}{drag coeffiecent (numeric). \code{Cd} can also be a function that returns
a drag coefficient for each wind speed, in which case a dynamic drag coefficient is used
is a function of wind speed that returns a vector of coefficients}

\item{rho}{density of air (kg m-3), standar for 20C @ 1 atm}
}
\description{
This method uses the 'profile' method Peltola 1999 to calculate wind
drag (Newtons) on tree segments using a vector of vertical distribution and wind profiel
}
\note{
There should be a note here
#test
}
\examples{
#there will be an exmaple here
}
