% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{garmsAndDean}
\alias{garmsAndDean}
\title{Critical turning moment for Longleaf pine}
\usage{
garmsAndDean(DBH_cm)
}
\arguments{
\item{DBH_cm}{tree diameter at breast height (cm)}
}
\description{
This function estimates estimates Critical turning moment based on tree size
(dbh) for longleaf pine. See Garms and Dean 2019 for details
Output is critical turning moment (Mmax) in kN-m
}
\examples{
#there will be an example here
}
