% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pred_LAI_longleaf}
\alias{pred_LAI_longleaf}
\title{Leaf area Index for Longleaf Pine}
\usage{
pred_LAI_longleaf(Dq, TPH, Hdom, age)
}
\arguments{
\item{Dq}{quadratic mean diameter (cm)}

\item{TPH}{tree density (trees per ha)}

\item{Hdom}{dominant tree height (m)}

\item{age}{stand age (years)}
}
\description{
This function calculates leaf area index for longleaf pine using equations from
Gonzalez-Benecke 2015
}
\note{
There should be a note here
}
\examples{
#there will be an example here
}
