% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wind_profile}
\alias{wind_profile}
\title{Create wind profile function for GALES}
\usage{
wind_profile(ws, LAI, ch, z = 1:40, plot = TRUE)
}
\arguments{
\item{ws}{wind speed (m/s) at canopy height}

\item{LAI}{Leaf area index (unitless) modifies wind speed}

\item{ch}{canopy height (m), canopy height at which ws measured}

\item{z}{height (m) at which to output wind speeds. Should be in 1 m increments}

\item{plot}{Logical; whether to print wind profile output to a plot}
}
\value{
a dataframe containing vector of windspeed \code{ws} (m/s) at each height \code{z}
}
\description{
This function generates a wind profile for modeling wind damage using the
modified Souza Model presented in de Santana et al. 2017.
}
\examples{
ws = wind_profile(ws=20, LAI=6, ch=30)
print(ws)
}
\keyword{Santana}
\keyword{Souza,}
\keyword{de}
\keyword{profile,}
\keyword{wind,}
